+++ packages/elements/src/calendar/__test__/calendar.range.test.js
@@ -84,5 +84,37 @@ describe('calendar/Range', () => {
       expect(el.values.join(','), 'range should populate through views').to.equal('2005-04-12,2005-05-02');
       expect(el).shadowDom.to.equalSnapshot();
     });
+    it('It should not be possible to deselect all range values', async () => {
+      const el = await fixture('<ef-calendar range view="2005-04" lang="en-GB"></ef-calendar>');
+      let cells = getDateCells(el);
+
+      cells[5].click(); // April 06
+      await elementUpdated(el);
+      expect(el.values.join(','), 'from should be populated').to.equal('2005-04-06');
+
+      cells[9].click(); // April 10
+      await elementUpdated(el);
+      expect(el.values.join(','), 'to should be populated').to.equal('2005-04-06,2005-04-10');
+
+      cells[9].click(); // April 10
+      await elementUpdated(el);
+      expect(el.values.join(','), 'from should be populated').to.equal('2005-04-10');
+
+      cells[5].click(); // April 06
+      await elementUpdated(el);
+      expect(el.values.join(','), 'from should be populated').to.equal('2005-04-06');
+    });
+    it('It should be possible to select the same value for From and To', async () => {
+      const el = await fixture('<ef-calendar range view="2005-04" lang="en-GB"></ef-calendar>');
+      let cells = getDateCells(el);
+
+      cells[5].click(); // April 06
+      await elementUpdated(el);
+      expect(el.values.join(','), 'from should be populated').to.equal('2005-04-06');
+
+      cells[5].click(); // April 06
+      await elementUpdated(el);
+      expect(el.values.join(','), 'should be able to select the same day as from').to.equal('2005-04-06,2005-04-06');
+    });
   });
 });