+++ packages/elements/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.8.8](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/elements@6.8.7...@refinitiv-ui/elements@6.8.8) (2023-06-30)
+
+
+### Bug Fixes
+
+* **number-field:** fix invalid regex ([#772](https://github.com/Refinitiv/refinitiv-ui/issues/772)) ([6af37fb](https://github.com/Refinitiv/refinitiv-ui/commit/6af37fb95bae9b194fdb3f4a6436b5a4a1048fca))
+* **select:** change role to be combobox ([#777](https://github.com/Refinitiv/refinitiv-ui/issues/777)) ([5a89914](https://github.com/Refinitiv/refinitiv-ui/commit/5a8991410bc0c52f9919cb009a93b0e895de1ede))
+
+
+
+
+
+## [6.8.7](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/elements@6.8.6...@refinitiv-ui/elements@6.8.7) (2023-06-23)
+
+
+### Bug Fixes
+
+* **password-field:** deregister overflow tooltip and text ellipsis ([#769](https://github.com/Refinitiv/refinitiv-ui/issues/769)) ([1579e08](https://github.com/Refinitiv/refinitiv-ui/commit/1579e08bdae5ecfa8709b293d31a29f5f479bcf2))
+
+
+
+
+
+## [6.8.6](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/elements@6.8.5...@refinitiv-ui/elements@6.8.6) (2023-06-20)
+
+
+### Bug Fixes
+
+* **calendar, datetime-picker:** prevent deselect date on calendar ([#735](https://github.com/Refinitiv/refinitiv-ui/issues/735)) ([4b0799a](https://github.com/Refinitiv/refinitiv-ui/commit/4b0799a19725cba456fa41be35de24b79df466c5))
+* **number-field:** fix Floating point Division on validate method ([#757](https://github.com/Refinitiv/refinitiv-ui/issues/757)) ([a7095ae](https://github.com/Refinitiv/refinitiv-ui/commit/a7095ae861e0948f95cc73ac68d31eb5585a54a6))
+
+
+
+
+
+## [6.8.5](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/elements@6.8.4...@refinitiv-ui/elements@6.8.5) (2023-06-08)
+
+**Note:** Version bump only for package @refinitiv-ui/elements
+
+
+
+
+
+## [6.8.4](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/elements@6.8.4-next.1...@refinitiv-ui/elements@6.8.4) (2023-06-07)
+
+**Note:** Version bump only for package @refinitiv-ui/elements
+
+
+
+
+
+## [6.8.4-next.1](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/elements@6.8.4-next.0...@refinitiv-ui/elements@6.8.4-next.1) (2023-06-06)
+
+**Note:** Version bump only for package @refinitiv-ui/elements
+
+
+
+
+
+## [6.8.3](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/elements@6.8.2...@refinitiv-ui/elements@6.8.3) (2023-06-06)
+
+**Note:** Version bump only for package @refinitiv-ui/elements
+
+
+
+
+
 ## [6.8.2](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/elements@6.8.1...@refinitiv-ui/elements@6.8.2) (2023-05-24)